pyproject.toml
[project]
name = "gym-sentiment-guard"
version = "0.1.0"
description = "Sentiment analysis tooling for gym reviews"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pandas>=2.1.0",
    "fasttext-wheel>=0.9.2",
    "typer>=0.12.5",
    "pyyaml>=6.0.1",
    "scikit-learn>=1.4.0",
    "matplotlib>=3.8.0",
    "seaborn>=0.13.0",
    "requests>=2.32.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.30.0",
    "pyarrow>=14.0.0",
    "numpy>=1.26.0",
    "joblib>=1.3.0",
    "pydantic>=2.5.0",
    "scipy>=1.11.0",
    "tenacity>=8.2.0",
    "transformers>=4.36.0",
    "torch>=2.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "ruff>=0.1.0",
]

[project.scripts]
gym = "gym_sentiment_guard.cli.main:app"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]

[tool.ruff]
target-version = "py311"
line-length = 100
exclude = [".venv","build","dist","data","notebooks/.ipynb_checkpoints"]

[tool.ruff.lint]
select = ["E","F","I","B","UP","N","SIM","PTH","S","T20"]
ignore = ["E501","S101","T201"]

[tool.ruff.lint.per-file-ignores]
"tests/**.py" = ["S101"]
"notebooks/**.py" = ["T201"]

[tool.ruff.format]
quote-style = "single"
indent-style = "space"
line-ending = "auto"
